Moved the frequency bin creation into the ColorFactory
